+++ packages/sdk/src/lib/dataProtectorCore/getGrantedAccess.ts
@@ -1,7 +1,8 @@
 import { WorkflowError, handleIfProtocolError } from '../../utils/errors.js';
 import { formatGrantedAccess } from '../../utils/format.js';
 import {
-  addressOrEnsOrAnySchema,
+  addressOrEnsSchema,
+  booleanSchema,
   numberBetweenSchema,
   positiveNumberSchema,
   throwIfMissing,
@@ -14,24 +15,25 @@ import { IExecConsumer } from '../types/internalTypes.js';
 
 export const getGrantedAccess = async ({
   iexec = throwIfMissing(),
-  protectedData = 'any',
-  authorizedApp = 'any',
-  authorizedUser = 'any',
+  protectedData,
+  authorizedApp,
+  authorizedUser,
+  isUserStrict = false,
   page,
   pageSize,
 }: IExecConsumer & GetGrantedAccessParams): Promise<GrantedAccessResponse> => {
-  const vProtectedData = addressOrEnsOrAnySchema()
-    .required()
+  const vProtectedData = addressOrEnsSchema()
     .label('protectedData')
     .validateSync(protectedData);
-  const vAuthorizedApp = addressOrEnsOrAnySchema()
-    .required()
+  const vAuthorizedApp = addressOrEnsSchema()
     .label('authorizedApp')
     .validateSync(authorizedApp);
-  const vAuthorizedUser = addressOrEnsOrAnySchema()
-    .required()
+  const vAuthorizedUser = addressOrEnsSchema()
     .label('authorizedUser')
     .validateSync(authorizedUser);
+  const vIsUserStrict = booleanSchema()
+    .label('isUserStrict')
+    .validateSync(isUserStrict);
   const vPage = positiveNumberSchema().label('page').validateSync(page);
   const vPageSize = numberBetweenSchema(10, 1000)
     .label('pageSize')
@@ -39,10 +41,12 @@ export const getGrantedAccess = async ({
 
   try {
     const { count, orders } = await iexec.orderbook.fetchDatasetOrderbook(
-      vProtectedData,
+      vProtectedData || 'any',
       {
-        app: vAuthorizedApp,
-        requester: vAuthorizedUser,
+        app: vAuthorizedApp || 'any',
+        requester: vAuthorizedUser || 'any',
+        isRequesterStrict: vIsUserStrict,
+        isAppStrict: true,
         page: vPage,
         pageSize: vPageSize,
       }