+++ packages/sdk/src/lib/types/coreTypes.ts
@@ -99,21 +99,28 @@ export type GetGrantedAccessParams = {
    * Default fetch for any protectedData
    */
 
-  protectedData?: AddressOrENS | 'any';
+  protectedData?: AddressOrENS;
 
   /**
    * Address or ENS of the app authorized to use the `protectedData`
    *
    * Default fetch for any app
    */
-  authorizedApp?: AddressOrENS | 'any';
+  authorizedApp?: AddressOrENS;
 
   /**
    * Address or ENS of the user authorized to use the `protectedData`
    *
    * Default fetch for any user
    */
-  authorizedUser?: AddressOrENS | 'any';
+  authorizedUser?: AddressOrENS;
+
+  /**
+   * Fetches the orderbook strictly specified for this user
+   *
+   * Default false for any user
+   */
+  isUserStrict?: boolean;
 
   /**
    * Index of the page to fetch
@@ -206,14 +213,14 @@ export type RevokeAllAccessParams = {
    *
    * Default revoke for any app
    */
-  authorizedApp?: AddressOrENS | 'any';
+  authorizedApp?: AddressOrENS;
 
   /**
    * Address or ENS of the user authorized to use the `protectedData`
    *
    * Default revoke for any user
    */
-  authorizedUser?: AddressOrENS | 'any';
+  authorizedUser?: AddressOrENS;
 
   /**
    * Callback function that will get called at each step of the process
@@ -286,7 +293,12 @@ export type ProcessProtectedDataParams = {
   /**
    * The workerpool to use for the application's execution. (default iExec production workerpool)
    */
-  workerpool?: AddressOrENS | 'any';
+  workerpool?: AddressOrENS;
+
+  /**
+   * Option to use a voucher for task payment (default: false)
+   */
+  useVoucher?: boolean;
 
   /**
    * Callback function that will get called at each step of the process